+++ libNOM.io/Meta/GameVersion.cs
@@ -42,8 +42,8 @@ internal static class GameVersion
     // EXTERNAL RELEASE: Add new game version.
     internal static GameVersionEnum Get(int baseVersion) => baseVersion switch
     {
-        // >= 4168 => GameVersionEnum.WorldsPartII, // 5.50
-        // >= 4167 => GameVersionEnum.TheCursedWithStarbornPhoenix, // 5.29
+        >= 4168 => GameVersionEnum.WorldsPartII, // 5.50
+        >= 4167 => GameVersionEnum.TheCursedWithStarbornPhoenix, // 5.29
         >= 4164 => GameVersionEnum.TheCursedWithCrossSave, // 5.25
         >= 4161 => GameVersionEnum.TheCursed, // 5.20
         >= 4155 => GameVersionEnum.Aquarius, // 5.10
@@ -82,6 +82,8 @@ internal static class GameVersion
     // EXTERNAL RELEASE: Add new game version.
     internal static GameVersionEnum Get(int baseVersion, JObject jsonObject) => baseVersion switch
     {
+        >= 4168 => GameVersionEnum.WorldsPartII, // 5.50
+        >= 4167 => GameVersionEnum.TheCursedWithStarbornPhoenix, // 5.29
         >= 4164 => GameVersionEnum.TheCursedWithCrossSave, // 5.25
         >= 4161 => GameVersionEnum.TheCursed, // 5.20
         >= 4155 => GameVersionEnum.Aquarius, // 5.10
@@ -116,6 +118,8 @@ internal static class GameVersion
     // EXTERNAL RELEASE: Add new game version.
     internal static GameVersionEnum Get(int baseVersion, string json) => baseVersion switch
     {
+        >= 4168 => GameVersionEnum.WorldsPartII, // 5.50
+        >= 4167 => GameVersionEnum.TheCursedWithStarbornPhoenix, // 5.29
         >= 4164 => GameVersionEnum.TheCursedWithCrossSave, // 5.25
         >= 4161 => GameVersionEnum.TheCursed, // 5.20
         >= 4155 => GameVersionEnum.Aquarius, // 5.10